app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
python-jose
python-multipart